requirements.txt
asgiref==3.3.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.1.7
django-embed-video==1.4.0
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.7
whitenoise==5.3.0
